i.atcorr: avoid float to double to float conversion, causing numerical instability
git-svn-id: https://svn.osgeo.org/grass/grass/trunk@73183 15284696-431f-4ddb-bdfa-cd5b030d7da7

+++ imagery/i.atcorr/6s.h
@@ -29,9 +29,9 @@ extern int init_6S(char* icnd_name);
   This requires lots of computations and therefore can be very
   time consuming.
 */
-extern void pre_compute_hv(const float height, const float vis);
-extern void pre_compute_v(const float vis);
-extern void pre_compute_h(const float height);
+extern void pre_compute_hv(const double height, const double vis);
+extern void pre_compute_v(const double vis);
+extern void pre_compute_h(const double height);
 
 struct TransformInput;
 /* Compute the input parameters used to do atmospheric correction on input values
